+++ src/components/ui/shadcn/calendar.tsx
@@ -4,8 +4,8 @@ import * as React from 'react';
 import { ChevronLeft, ChevronRight } from 'lucide-react';
 import { DayPicker } from 'react-day-picker';
 
-import { cn } from '@/lib/utils';
-import { buttonVariants } from '@/components/ui/button';
+import { cn } from '@/utils/lib/cn';
+import { buttonVariants } from '@/components/ui/shadcn/button';
 
 export type CalendarProps = React.ComponentProps<typeof DayPicker>;
