+++ src/components/ui/shadcn/popover.tsx
@@ -3,7 +3,7 @@
 import * as React from 'react';
 import * as PopoverPrimitive from '@radix-ui/react-popover';
 
-import { cn } from '@/lib/utils';
+import { cn } from '@/utils/lib/cn';
 
 const Popover = PopoverPrimitive.Root;
